Rename sol_bpf.h to solana_sdk.h
