+++ build.rs
@@ -25,7 +25,7 @@ fn main() {
             + &"/bpf".to_string();
 
         println!("cargo:rerun-if-changed=programs/bpf/c/sdk/bpf.mk");
-        println!("cargo:rerun-if-changed=programs/bpf/c/sdk/inc/sol_bpf.h");
+        println!("cargo:rerun-if-changed=programs/bpf/c/sdk/inc/solana_sdk.h");
         println!("cargo:rerun-if-changed=programs/bpf/c/makefile");
         println!("cargo:rerun-if-changed=programs/bpf/c/src/move_funds.c");
         println!("cargo:rerun-if-changed=programs/bpf/c/src/noop.c");
