+++ programs/bpf/c/src/move_funds.c
@@ -3,7 +3,7 @@
  * another
  */
 
-#include <sol_bpf.h>
+#include <solana_sdk.h>
 
 /**
  * Number of SolKeyedAccounts expected. The program should bail if an
